refactor receipt components
